Need to create product data pages

+++ engine-server/db/migrate/20171104154209_create_products.rb
@@ -4,9 +4,9 @@ class CreateProducts < ActiveRecord::Migration[5.1]
       t.string :name
       t.string :base_currency
       t.string :quote_currency
-      t.decimal :base_min_size
+      t.float :base_min_size
       t.integer :base_max_size
-      t.decimal :quote_increment
+      t.float :quote_increment
       t.string :display_name
       t.boolean :margin_enabled
